requirements.txt
# Core dependencies
python-dotenv>=1.0.0

# API clients
openai>=1.0.0
pinecone>=5.0.0

# UI
streamlit>=1.28.0

# Testing (not needed in production but included for CI)
pytest>=7.0.0
